fix(db): schema bugs — quantity decimal + status width
  - G-1: product_units.status VARCHAR(20)→VARCHAR(30) — tránh truncate RETURNED_TO_SUPPLIER (21 ký tự)

  - G-2: import_receipt_items.quantity INT→DECIMAL(15,2) — hỗ trợ bulk (mét/kg)

  - G-3: export_receipt_items.quantity INT→DECIMAL(15,2) — hỗ trợ xuất bulk lẻ (US-33)

  - Entities: Integer→BigDecimal, requests/responses đồng bộ

  - Service: ImportReceiptService, ExportReceiptService dùng BigDecimal cho quantity

  - V8 migration: ALTER TABLE cho 3 cột

+++ backend/src/main/java/org/dawn/backend/controller/inventory/request/ExportReceiptRequest.java
@@ -11,7 +11,7 @@ public record ExportReceiptRequest(
 ) {
     public record ExportItemRequest(
             Long productId,
-            Integer quantity,
+            BigDecimal quantity,
             BigDecimal unitPrice
     ) {}
 }
